Refactor move script command to proper files
