+++ ShakerGenerator/Script/Commands.cpp
@@ -1,0 +1,32 @@
+#include "Commands.h"
+
+std::map<std::string, std::pair<unsigned int, unsigned int>> CommandKeyOutput::Escape_map_ = {
+   {"\\(ESC)", { 8, 2 }},
+   {"\\(TAB)", { 8, 4 }},
+   {"\\(CAP)", { 8, 6 }},
+   {"\\(SHI)", { 2, 5 }},
+   {"\\(CTR)", { 2, 7 }},
+   {"\\(COP)", { 2, 1 }},
+   {"\\(CLR)", { 2, 0 }},
+   {"\\(DEL)", { 9, 7 }},
+   {"\\(RET)", { 2, 2 }},
+   {"\\(ENT)", { 0, 6 }},
+   {"\\(ARL)", { 1, 0 }},
+   {"\\(ARR)", { 0, 1 }},
+   {"\\(ARU)", { 0, 0 }},
+   {"\\(ARD)", { 0, 2 }},
+   {"\\(F0)", { 1, 7 }},
+   {"\\(F1)", { 1, 5 }},
+   {"\\(F2)", { 1, 6 }},
+   {"\\(F3)", { 0, 5 }},
+   {"\\(F4)", { 2, 4 }},
+   {"\\(F5)", { 1, 4 }},
+   {"\\(F6)", { 0, 4 }},
+   {"\\(F7)", { 1, 2 }},
+   {"\\(F8)", { 1, 3 }},
+   {"\\(F9)", { 0, 3 }},
+   {"\\({)", { 2, 1 }},
+   {"\\(})", { 2, 3 }},
+   {"\\(\\)", { 2, 6 }},
+   {"\\(')", { 5, 1 }}
+};
