+++ ShakerGenerator/TestUtils.cpp
@@ -20,37 +20,7 @@
 #include <map>
 #include <string>
 #include "TestUtils.h"
-
-std::map<std::string, std::pair<unsigned int, unsigned int>> Escape_map_ = {
-   {"\\(ESC)", { 8, 2 }},
-   {"\\(TAB)", { 8, 4 }},
-   {"\\(CAP)", { 8, 6 }},
-   {"\\(SHI)", { 2, 5 }},
-   {"\\(CTR)", { 2, 7 }},
-   {"\\(COP)", { 2, 1 }},
-   {"\\(CLR)", { 2, 0 }},
-   {"\\(DEL)", { 9, 7 }},
-   {"\\(RET)", { 2, 2 }},
-   {"\\(ENT)", { 0, 6 }},
-   {"\\(ARL)", { 1, 0 }},
-   {"\\(ARR)", { 0, 1 }},
-   {"\\(ARU)", { 0, 0 }},
-   {"\\(ARD)", { 0, 2 }},
-   {"\\(F0)", { 1, 7 }},
-   {"\\(F1)", { 1, 5 }},
-   {"\\(F2)", { 1, 6 }},
-   {"\\(F3)", { 0, 5 }},
-   {"\\(F4)", { 2, 4 }},
-   {"\\(F5)", { 1, 4 }},
-   {"\\(F6)", { 0, 4 }},
-   {"\\(F7)", { 1, 2 }},
-   {"\\(F8)", { 1, 3 }},
-   {"\\(F9)", { 0, 3 }},
-   {"\\({)", { 2, 1 }},
-   {"\\(})", { 2, 3 }},
-   {"\\(\\)", { 2, 6 }},
-   {"\\(')", { 5, 1 }}
-};
+#include "Script/CSLScriptRunner.h"
 
 FileLog::FileLog(const char* file)
 {
@@ -328,7 +298,7 @@ void TestDump::SetScreenshotHandler()
     }
 }
 
-bool TestDump::Test(std::filesystem::path conf, std::filesystem::path initfile, CommandList* cmd_list, bool bFixedSpeed, int seed)
+bool TestDump::Test(std::filesystem::path conf, std::filesystem::path initfile, std::filesystem::path scriptfile, bool bFixedSpeed, int seed)
 {
    // Creation dela machine
    
@@ -354,13 +324,11 @@ bool TestDump::Test(std::filesystem::path conf, std::filesystem::path initfile, 
    srand(seed);
    machine_->SetFixedSpeed(bFixedSpeed);
 
-   // Run preliminary actions
-   bool no_error = cmd_list->RunFirstCommand(machine_);
-   while (cmd_list->IsFinished() == false && no_error)
-   {
-      no_error = cmd_list->RunNextCommand(machine_);
-   }
-   return no_error;
+   CSLScriptRunner runner(machine_);
+
+   runner.LoadScript(scriptfile);
+   
+   return runner.Run();
 }
